+++ app/src/main/java/com/numero/material_gallery/MaterialGalleryApplication.kt
@@ -1,8 +1,8 @@
 package com.numero.material_gallery
 
 import android.app.Application
+import com.numero.material_gallery.repository.ConfigRepositoryImpl
 import com.numero.material_gallery.repository.ConfigRepository
-import com.numero.material_gallery.repository.IConfigRepository
 import org.koin.android.ext.android.get
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
@@ -16,11 +16,11 @@ class MaterialGalleryApplication : Application() {
             androidContext(this@MaterialGalleryApplication)
             modules(appModule)
         }
-        val configRepository: IConfigRepository = get()
+        val configRepository: ConfigRepository = get()
         configRepository.getCurrentTheme().apply()
     }
 
     private val appModule = module {
-        single { ConfigRepository(androidContext()) as IConfigRepository }
+        single { ConfigRepositoryImpl(androidContext()) as ConfigRepository }
     }
 }
