+++ app/src/main/java/com/numero/material_gallery/activity/BottomSheetActivity.kt
@@ -12,13 +12,13 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_bottom_sheet.*
 import org.koin.android.ext.android.inject
 
 class BottomSheetActivity : AppCompatActivity(R.layout.activity_bottom_sheet) {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     private lateinit var behavior: BottomSheetBehavior<LinearLayout>
