+++ app/src/main/java/com/numero/material_gallery/activity/CheckboxActivity.kt
@@ -8,13 +8,13 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_checkbox.*
 import org.koin.android.ext.android.inject
 
 class CheckboxActivity : AppCompatActivity(R.layout.activity_checkbox) {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
