+++ app/src/main/java/com/numero/material_gallery/activity/ExtendedFabActivity.kt
@@ -9,13 +9,13 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_extended_fab.*
 import org.koin.android.ext.android.inject
 
 class ExtendedFabActivity : AppCompatActivity(R.layout.activity_extended_fab) {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
