+++ app/src/main/java/com/numero/material_gallery/activity/NavigationDrawerActivity.kt
@@ -12,14 +12,14 @@ import androidx.core.view.GravityCompat
 import com.google.android.material.navigation.NavigationView
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_navigation_drawer.*
 import kotlinx.android.synthetic.main.app_bar_navigation_drawer.*
 import org.koin.android.ext.android.inject
 
 class NavigationDrawerActivity : AppCompatActivity(R.layout.activity_navigation_drawer), NavigationView.OnNavigationItemSelectedListener {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
