+++ app/src/main/java/com/numero/material_gallery/activity/bottom_app_bar/BottomAppBarActivity.kt
@@ -10,13 +10,13 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.bottomappbar.BottomAppBar
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_bottom_app_bar.*
 import org.koin.android.ext.android.inject
 
 class BottomAppBarActivity : AppCompatActivity(R.layout.activity_bottom_app_bar) {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
