+++ app/src/main/java/com/numero/material_gallery/activity/top_app_bar/CollapsingActivity.kt
@@ -8,13 +8,13 @@ import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import com.numero.material_gallery.R
 import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
-import com.numero.material_gallery.repository.IConfigRepository
+import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_collapsing.*
 import org.koin.android.ext.android.inject
 
 class CollapsingActivity : AppCompatActivity(R.layout.activity_collapsing) {
 
-    private val configRepository by inject<IConfigRepository>()
+    private val configRepository by inject<ConfigRepository>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(configRepository.themeRes)
